Added CRON for daily stats generation
